client/requirements.txt
Flask
flask-login
requests
pydantic
python-dateutil
jsonpath2